Refactored old "TestAux". Added MFEMHesthaven Comparison.
